+++ interactions/api/models/guild.py
@@ -313,6 +313,8 @@ class Guild(DictSerializerMixin):
         :param delete_message_days?: Number of days to delete messages, from 0 to 7. Defaults to 0
         :type delete_message_days: Optional[int]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.create_guild_ban(
             guild_id=int(self.id),
             user_id=member_id,
@@ -332,6 +334,8 @@ class Guild(DictSerializerMixin):
         :param reason?: The reason for the removal of the ban
         :type reason: Optional[str]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.remove_guild_ban(
             guild_id=int(self.id),
             user_id=user_id,
@@ -350,6 +354,8 @@ class Guild(DictSerializerMixin):
         :param reason?: The reason for the kick
         :type reason: Optional[str]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.create_guild_kick(
             guild_id=int(self.id),
             user_id=member_id,
@@ -371,6 +377,8 @@ class Guild(DictSerializerMixin):
         :param reason?: The reason why the roles are added
         :type reason: Optional[str]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if isinstance(role, Role):
             await self._client.add_member_role(
                 guild_id=int(self.id),
@@ -401,6 +409,8 @@ class Guild(DictSerializerMixin):
         :param reason?: The reason why the roles are removed
         :type reason: Optional[str]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if isinstance(role, Role):
             await self._client.remove_member_role(
                 guild_id=int(self.id),
@@ -442,7 +452,8 @@ class Guild(DictSerializerMixin):
         :return: The created Role
         :rtype: Role
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         payload = Role(
             name=name,
             color=color,
@@ -467,6 +478,8 @@ class Guild(DictSerializerMixin):
         :return: The member searched for
         :rtype: Member
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         res = await self._client.get_member(
             guild_id=int(self.id),
             member_id=member_id,
@@ -482,6 +495,8 @@ class Guild(DictSerializerMixin):
         :param channel_id: The id of the channel to delete
         :type channel_id: int
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.delete_channel(
             channel_id=channel_id,
         )
@@ -498,7 +513,8 @@ class Guild(DictSerializerMixin):
         :param reason?: The reason of the deletion
         :type reason: Optional[str]
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.delete_guild_role(
             guild_id=int(self.id),
             role_id=role_id,
@@ -534,7 +550,8 @@ class Guild(DictSerializerMixin):
         :return: The modified role object
         :rtype: Role
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         roles = await self._client.get_all_roles(guild_id=int(self.id))
         for i in roles:
             if int(i["id"]) == role_id:
@@ -594,7 +611,8 @@ class Guild(DictSerializerMixin):
         :return: The created channel
         :rtype: Channel
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if type in [
             ChannelType.DM,
             ChannelType.DM.value,
@@ -664,6 +682,8 @@ class Guild(DictSerializerMixin):
         :return: The modified channel
         :rtype: Channel
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         ch = Channel(**await self._client.get_channel(channel_id=channel_id))
 
         _name = ch.name if not name else name
@@ -730,7 +750,8 @@ class Guild(DictSerializerMixin):
         :return: The modified member
         :rtype: Member
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         payload = {}
         if nick:
             payload["nick"] = nick
@@ -760,10 +781,14 @@ class Guild(DictSerializerMixin):
 
     async def get_preview(self) -> "GuildPreview":
         """Get the guild's preview."""
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         return GuildPreview(**await self._client.get_guild_preview(guild_id=int(self.id)))
 
     async def leave(self) -> None:
         """Removes the bot from the guild."""
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.leave_guild(guild_id=int(self.id))
 
     async def modify(
@@ -833,7 +858,8 @@ class Guild(DictSerializerMixin):
         :return: The modified guild
         :rtype: Guild
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if (
             suppress_join_notifications is None
             and suppress_premium_subscriptions is None
@@ -926,7 +952,8 @@ class Guild(DictSerializerMixin):
         :return: The created event
         :rtype: ScheduledEvents
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if entity_type != EntityType.EXTERNAL and not channel_id:
             raise ValueError(
                 "channel_id is required when entity_type is not external!"
@@ -994,7 +1021,8 @@ class Guild(DictSerializerMixin):
         :return: The modified event
         :rtype: ScheduledEvents
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         if entity_type == EntityType.EXTERNAL and not entity_metadata:
             raise ValueError(
                 "entity_metadata is required for external events!"
@@ -1037,6 +1065,8 @@ class Guild(DictSerializerMixin):
         :param event_id: The id of the event to delete
         :type event_id: int
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         await self._client.delete_scheduled_event(
             guild_id=self.id,
             guild_scheduled_event_id=Snowflake(event_id),
@@ -1049,6 +1079,8 @@ class Guild(DictSerializerMixin):
         :return: The channels of the guild.
         :rtype: List[Channel]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         res = self._client.get_all_channels(int(self.id))
         channels = [Channel(**channel, _client=self._client) for channel in res]
         return channels
@@ -1060,6 +1092,8 @@ class Guild(DictSerializerMixin):
         :return: The roles of the guild.
         :rtype: List[Role]
         """
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         res = self._client.get_all_roles(int(self.id))
         roles = [Role(**role, _client=self._client) for role in res]
         return roles
@@ -1082,7 +1116,8 @@ class Guild(DictSerializerMixin):
         :return: List of guild roles with updated hierarchy
         :rtype: List[Role]
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         _role_id = role_id.id if isinstance(role_id, Role) else role_id
         res = await self._client.modify_guild_role_position(
             guild_id=int(self.id), position=position, role_id=_role_id, reason=reason
@@ -1097,7 +1132,8 @@ class Guild(DictSerializerMixin):
         :return: List of banned users with reasons
         :rtype: List[dict]
         """
-
+        if not self._client:
+            raise AttributeError("HTTTPClient not found!")
         res = await self._client.get_guild_bans(int(self.id))
         for ban in res:
             ban["user"] = User(**ban["user"])
